+++ src/components/EditSections/EditUserRoles/helpers/index.js
@@ -1,0 +1,3 @@
+// eslint-disable-next-line import/prefer-default-export
+export { default as getInitialFiltersState } from './getInitialFiltersState';
+export { default as filtersConfig } from './filtersConfig';