+++ src/routes/UserRecordContainer.js
@@ -22,8 +22,8 @@ class UserRecordContainer extends React.Component {
       path: 'users/:{id}',
       clear: false,
       shouldRefresh: (resource, action, refresh) => {
-        const { path } = action.meta;
-        return refresh || (path && path.match(/link/));
+        const { path, name } = action.meta;
+        return (name && name !== 'resetPassword') && (refresh || (path && path.match(/link/)));
       },
       throwErrors: false,
     },
@@ -127,9 +127,19 @@ class UserRecordContainer extends React.Component {
       type: 'okapi',
       throwErrors: false,
       POST: {
-        path: 'perms/users',
+        path: (queryParams, pathComponents, resourceData, config, props) => {
+          if (props.stripes.hasInterface('roles')) {
+            return undefined;
+          }
+          return 'perms/users';
+        }
       },
-      path: 'perms/users/:{id}',
+      path: (queryParams, pathComponents, resourceData, config, props) => {
+        if (props.stripes.hasInterface('roles')) {
+          return undefined;
+        }
+        return `perms/users/${pathComponents.id}`;
+      },
       params: { full: 'true', indexField: 'userId' },
     },
     // NOTE: 'indexField', used as a parameter in the userPermissions paths,
@@ -143,22 +153,47 @@ class UserRecordContainer extends React.Component {
       resourceShouldRefresh: true,
       DELETE: {
         pk: 'permissionName',
-        path: 'perms/users/:{id}/permissions',
+        path: (queryParams, pathComponents, resourceData, config, props) => {
+          if (props.stripes.hasInterface('roles')) {
+            return undefined;
+          }
+          return `perms/users/${pathComponents.id}/permissions`;
+        },
         params: { indexField: 'userId' },
       },
       GET: {
-        path: 'perms/users/:{id}/permissions',
+        path: (queryParams, pathComponents, resourceData, config, props) => {
+          if (props.stripes.hasInterface('roles')) {
+            return undefined;
+          }
+          return `perms/users/${pathComponents.id}/permissions`;
+        },
         params: { full: 'true', indexField: 'userId' },
       },
       PUT: {
-        path: 'perms/users/%{permUserId}',
+        path: (queryParams, pathComponents, resourceData, config, props) => {
+          if (props.stripes.hasInterface('roles')) {
+            return undefined;
+          }
+          return 'perms/users';
+        },
       },
-      path: 'perms/users/:{id}/permissions',
+      path: (queryParams, pathComponents, resourceData, config, props) => {
+        if (props.stripes.hasInterface('roles')) {
+          return undefined;
+        }
+        return `perms/users/${pathComponents.id}/permissions`;
+      },
       params: { indexField: 'userId' },
     },
     settings: {
       type: 'okapi',
-      path: 'users/configurations/entry',
+      path: (queryParams, pathComponents, resourceData, config, props) => {
+        if (props.stripes.hasInterface('users', '16.1')) {
+          return 'users/configurations/entry';
+        }
+        return null;
+      },
     },
     requestPreferences: {
       type: 'okapi',