+++ src/views/UserDetail/UserDetail.js
@@ -44,6 +44,7 @@ import {
   ProxyPermissions,
   PatronBlock,
   UserPermissions,
+  UserRoles,
   UserLoans,
   UserRequests,
   UserAccounts,
@@ -200,6 +201,7 @@ class UserDetail extends React.Component {
         requestsSection: false,
         accountsSection: false,
         permissionsSection: false,
+        rolesSection: false,
         servicePointsSection: false,
         notesAccordion: false,
         customFields: false,
@@ -218,6 +220,19 @@ class UserDetail extends React.Component {
   componentWillUnmount() {
     this._isMounted = false;
   }
+
+  /**
+   * showPermissionsAccordion
+   * Return true unless the `roles` interface is present; then return false.
+   * When `roles` is present, this indicates access management is handled
+   * by keycloak; thus, roles, policies, and capabilites are used to manage
+   * access, not the legacy permissions system.
+   *
+   * @returns boolean true unless the `roles` interface is present
+   */
+  showPermissionsAccordion = () => {
+    return !this.props.stripes.hasInterface('roles');
+  };
 
   getUser = () => {
     const { resources, match: { params: { id } } } = this.props;
@@ -867,17 +882,29 @@ class UserDetail extends React.Component {
                   )
                 }
 
-                <IfPermission perm="perms.users.get">
-                  <IfInterface name="permissions" version="5.0">
-                    <UserPermissions
-                      expanded={sections.permissionsSection}
-                      onToggle={this.handleSectionToggle}
-                      accordionId="permissionsSection"
-                      user={this.getUser()}
-                      {...this.props}
-                    />
-                  </IfInterface>
-                </IfPermission>
+                { this.showPermissionsAccordion() &&
+                  <IfPermission perm="perms.users.get">
+                    <IfInterface name="permissions" version="5.0">
+                      <UserPermissions
+                        expanded={sections.permissionsSection}
+                        onToggle={this.handleSectionToggle}
+                        accordionId="permissionsSection"
+                        user={user}
+                        {...this.props}
+                      />
+                    </IfInterface>
+                  </IfPermission>
+                }
+
+                { !this.showPermissionsAccordion() &&
+                  <UserRoles
+                    expanded={sections.rolesSection}
+                    onToggle={this.handleSectionToggle}
+                    accordionId="rolesSection"
+                    user={user}
+                    {...this.props}
+                  />
+                }
 
                 <IfPermission perm="inventory-storage.service-points.collection.get,inventory-storage.service-points-users.collection.get">
                   <IfInterface name="service-points-users" version="1.0">